backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Groq API
groq>=0.4.0

# Vector Database
chromadb>=0.4.22

# Text Processing
tiktoken>=0.5.2

# PDF Processing
PyPDF2>=3.0.0

# HTTP Client
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.1